+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2026-06-02]
+### Added
+- Added support for AFC correctly working with Snapmaker U1 pause/resume and power loss resume functionality.
+
 ## [2026-05-28]
 ### Added
 - Added support for AFC working on Snapmaker U1